check for empty fileds

+++ app/src/main/java/com/example/headsUpRoom/EditAndDelete.kt
@@ -31,29 +31,41 @@ class EditAndDelete : AppCompatActivity() {
         val btBack = findViewById<Button>(R.id.btBack1)
 
         btedit.setOnClickListener {
-            val ID = id.text.toString().toInt()
-            val s = CelebrityDetails(ID,Nname.text.toString(),Ntaboo1.text.toString(),Ntaboo2.text.toString(),Ntaboo3.text.toString())
-            CoroutineScope(Dispatchers.IO).launch {
-                CelebrityDatabase.getInstance(applicationContext).CelebrityDao().updateCelebrity(s)
+            if (id.text.isEmpty())
+                Toast.makeText(this, "Enter the id please", Toast.LENGTH_SHORT).show()
+            else {
+                val ID = id.text.toString().toInt()
+                if(Nname.text.isEmpty()||Ntaboo1.text.isEmpty()||Ntaboo2.text.isEmpty()||Ntaboo3.text.isEmpty())
+                    Toast.makeText(this, "Fill all fields please", Toast.LENGTH_SHORT).show()
+                else {
+                    val s = CelebrityDetails(ID, Nname.text.toString(), Ntaboo1.text.toString(), Ntaboo2.text.toString(), Ntaboo3.text.toString())
+                    CoroutineScope(Dispatchers.IO).launch {
+                        CelebrityDatabase.getInstance(applicationContext).CelebrityDao().updateCelebrity(s)
+                    }
+                    Toast.makeText(applicationContext, "data updated successfully! ", Toast.LENGTH_SHORT)
+                            .show()
+                    println("updated item")
+                    id.setText("")
+                    Nname.setText("")
+                    Ntaboo1.setText("")
+                    Ntaboo2.setText("")
+                    Ntaboo3.setText("")
+                }
             }
-            Toast.makeText(applicationContext, "data updated successfully! ", Toast.LENGTH_SHORT)
-                .show()
-            println("updated item")
-            id.setText("")
-            Nname.setText("")
-            Ntaboo1.setText("")
-            Ntaboo2.setText("")
-            Ntaboo3.setText("")
         }
 
         btdel.setOnClickListener {
-            val del = CelebrityDetails(DelID.text.toString().toInt(),"","","","")
-            CoroutineScope(Dispatchers.IO).launch {
-                CelebrityDatabase.getInstance(applicationContext).CelebrityDao().deleteCelebrity(del)
+            if (DelID.text.isEmpty())
+                Toast.makeText(this, "Enter the id please", Toast.LENGTH_SHORT).show()
+            else {
+                val del = CelebrityDetails(DelID.text.toString().toInt(), "", "", "", "")
+                CoroutineScope(Dispatchers.IO).launch {
+                    CelebrityDatabase.getInstance(applicationContext).CelebrityDao().deleteCelebrity(del)
+                }
+                Toast.makeText(applicationContext, "data deleted successfully! ", Toast.LENGTH_SHORT)
+                        .show()
+                DelID.setText("")
             }
-            Toast.makeText(applicationContext, "data deleted successfully! ", Toast.LENGTH_SHORT)
-                .show()
-            DelID.setText("")
         }
         btBack.setOnClickListener {
             intent = Intent(applicationContext, AddCelebrity::class.java)
